requirements.txt
flipside==2.0.8
pandas==1.4.3
python-dotenv==0.21.1
pymongo==4.6.0
solana==0.30.2